fix: rank 타입을 int로 수정

+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/reading/dto/response/MonthlyReadingRankResponse.java
@@ -12,7 +12,7 @@ public record MonthlyReadingRankResponse(
         String nickname,
 
         @Schema(type = "integer", format = "int32", required = true)
-        long rank,
+        int rank,
 
         @Schema(type = "integer", format = "int32", required = true)
         int readCount
